Adjusting TableEqual to Testing Lib and Bugfix in HashJoinProbe
The TableEquality was now moved to libtesting to avoid dependencies!!!
of gtest for the storage library. This required lots of changes in the
makefile because now libtesting requires libstorage and libftprinter-hyr
etc. This might not be optimal, but better than keeping this in libstorage.

The bug in HashJoinProbe was a not initialized variable for the selfjoin.