Radix Clustered Hash Join is now parallel.
The radix join adds the following plan operatoins:

- Histogram - calculates occurrences per value in a table
- Histogram 2nd pass
- Radix Cluster - clusters an input table based on a prefix sum list and
  a hash value
- Radix Cluster 2nd Pass
- Prefix Sum - generates a prefix sum array based on the histogram
- Merge Prefix Sum
- Barrier
- Create Radix Table

The script in tools/joiner.rb can be used to generate paralle
radix join plans. In addition there are a few changes to the
interface of the fixed length vector to better handle the
down-casted data access. This includes a inc() and
atomic_inc() method to increment a certain value at poin.

Limitations:
  -> Warning: Only Integer values can be joined