JsonTable Plan Operation
Adding the possibility to specify tables directly from JSON without
specifying the header. Supports the following notation:

	"0" : {
  		"type" : "JsonTable",
  		"names" : ["A", "B", "C"],
  		"types" : ["INTEGER", "STRING", "FLOAT"],
  		"groups" : [3],
  		"useStore" : true
  	}

Follows the same properties as the TableBuilder class that is used below. Furthermore,
it has the useStore flag to indicate if the result should be a Store class with the
default Merge set.

The plan operation now supports specifying a data section
that allows to directly submit intput data to the plan
operation.

    {
        "operators": {
            "-1" : {
                "type": "TableLoad",
                "table": "reference",
                "filename" : "tables/revenue.tbl"
            },
            "0": {
                "type": "JsonTable",
                "names": ["year", "quarter", "amount"],
                "types" : ["INTEGER", "INTEGER", "INTEGER"],
                "groups" : [1,1,1],
                "data" : [
                    ["2009","1","2000"],
                    ["2009","2","2500"],
                    ["2009","3","3000"],
                    ["2009","4","4000"],
                    ["2010","1","2400"],
                    ["2010","2","2800"],
                    ["2010","3","3200"],
                    ["2010","4","3600"]
                ]
            }
        },
        "edges" : [["0", "0"]]
    }

The above being an example from the unit tests. Data must be given as strings
regardless of the types and will be casted correctly later