Adding Test Uncommited Helper
This is a small helper script that will only run those unit tests with files
that were modified since the last push.

make test_uncommited

executes this command. This idea is more or less take from the rails
test:uncommited rake command. The module checking is done in the small
python script that maps the source paths agains the current module list.