Fix test output on non-NUMA systems
The group committer will try to bind to the current NUMA-node no matter what.
This will fail on non-NUMA systems (read: the vagrant VM). Therefore, `make
test` will have a message on stderr intermingled in its output even if the test
is not failing. This is ugly.

I propose to fix this by only attempting to bind to a node, if there are
actually different nodes on the system. I found this to be the cleanest
solution.