Fixed HashJoin adding already hashed value to HashMap
"It appears as if this is indeed a bug. My first impression is that `extract<T>` in ` getGroupKey` should return the value id in both cases, not only for the selfjoin. The actual hashing is done by the GroupKeyHash, which is passed as the third template parameter to the unordered_map. As such, there is no need to also hash the key."